[Update] Training layouts updated Note: Creating a training must have multiple layouts (2 at the moment). Java code still to be implemented

+++ app/src/main/java/studentcompany/sportgest/Trainings/TrainingListActivity.java
@@ -15,17 +15,24 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import studentcompany.sportgest.R;
+import studentcompany.sportgest.daos.Pair;
 import studentcompany.sportgest.daos.Training_DAO;
+import studentcompany.sportgest.daos.Training_Exercise_DAO;
 import studentcompany.sportgest.daos.exceptions.GenericDAOException;
+import studentcompany.sportgest.domains.Exercise;
 import studentcompany.sportgest.domains.Training;
+import studentcompany.sportgest.domains.TrainingExercise;
 
 public class TrainingListActivity extends AppCompatActivity implements ListTraining_Fragment.OnItemSelected  {
 
     private Training_DAO training_dao;
+    private Training_Exercise_DAO training_exercise_dao;
     private List<Training> trainingList;
+    private List<TrainingExercise> trainingExercisesList;
     private int currentPos = -1;
     private Menu mOptionsMenu;
 
@@ -40,9 +47,10 @@ public class TrainingListActivity extends AppCompatActivity implements ListTrain
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_training_list);
 
-
         try {
             training_dao = new Training_DAO(getApplicationContext());
+            training_exercise_dao = new Training_Exercise_DAO(getApplicationContext());
+
             trainingList = training_dao.getAll();
             if(trainingList.isEmpty()) {
                 new TrainingTestData(getApplicationContext());
@@ -72,7 +80,24 @@ public class TrainingListActivity extends AppCompatActivity implements ListTrain
 
         for(Training t: trainingList)
             list.add(t.getTitle());
+        Collections.sort(list);
+        return list;
+    }
 
+    public List<String> getExercisesNamesList(List<TrainingExercise> exerciseList){
+        ArrayList<String> list = new ArrayList<>();
+        Exercise e;
+
+        for(TrainingExercise te: exerciseList) {
+            if (te != null) {
+                e = te.getExercise();
+                if(e != null) {
+                    list.add(e.getTitle() + " x" + te.getRepetitions());
+                }
+            }
+            e=null;
+        }
+        Collections.sort(list);
         return list;
     }
 
@@ -80,12 +105,27 @@ public class TrainingListActivity extends AppCompatActivity implements ListTrain
         mDetailsTraining.clearDetails();
         mListTrainings.removeItem(currentPos);
 
-        training_dao.deleteById(trainingList.get(currentPos).getId());
+        try {
+            Training training = training_dao.getById(trainingList.get(currentPos).getId());
+            if(training != null) {
+                //search by training id
+                TrainingExercise trainingExercise = new TrainingExercise(-1, training, null, -1);
+                //remove list of exercises
+                ArrayList<TrainingExercise> previousTrainingExercises = (ArrayList) training_exercise_dao.getByCriteria(trainingExercise);
+                for (TrainingExercise te : previousTrainingExercises) {
+                    training_exercise_dao.delete(te);
+                }
+                //remove training
+                training_dao.deleteById(trainingList.get(currentPos).getId());
+            }
+        }catch (GenericDAOException ex){
+            ex.printStackTrace();
+        }
         trainingList.remove(currentPos);
 
         currentPos = -1;
-        MenuItem item = mOptionsMenu.findItem(R.id.action_del);
-        item.setVisible(false);
+        mOptionsMenu.findItem(R.id.Delete).setVisible(false);
+        mOptionsMenu.findItem(R.id.Edit).setVisible(false);
     }
     /************************************
      ****     Listener Functions     ****
@@ -96,12 +136,19 @@ public class TrainingListActivity extends AppCompatActivity implements ListTrain
 
         if(training != null){
             if(currentPos == -1) {
-                MenuItem item = mOptionsMenu.findItem(R.id.action_del);
-                item.setVisible(true);
+                mOptionsMenu.findItem(R.id.Delete).setVisible(true);
+                mOptionsMenu.findItem(R.id.Edit).setVisible(true);
             }
 
             currentPos = position;
-            mDetailsTraining.showTraining(training);
+            try {
+                TrainingExercise trainingExercise = new TrainingExercise(-1, training, null, -1);
+                trainingExercisesList = training_exercise_dao.getByCriteria(trainingExercise);
+            } catch (GenericDAOException ex){
+                ex.printStackTrace();
+                trainingExercisesList = new ArrayList<>();
+            }
+            mDetailsTraining.showTraining(training, getExercisesNamesList(trainingExercisesList));
         }
     }
 
@@ -152,27 +199,60 @@ public class TrainingListActivity extends AppCompatActivity implements ListTrain
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_toolbar_crud, menu);
         mOptionsMenu = menu;
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menu_users_view, menu);
+        menu.findItem(R.id.Edit).setVisible(false);
+        menu.findItem(R.id.Delete).setVisible(false);
+        menu.findItem(R.id.Save).setVisible(false);
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
+        Intent intent;
         // Handle item selection
         switch (item.getItemId()) {
-            case R.id.action_add:
-                Intent intent = new Intent(this, CreateTrainingActivity.class);
-                startActivity(intent);
+            case R.id.Add:
+                intent = new Intent(this, CreateTrainingActivity.class);
+                startActivityForResult(intent, 0);
                 return true;
 
-            case R.id.action_del:
+            case R.id.Edit:
+                intent = new Intent(this, CreateTrainingActivity.class);
+                //put current training ID in extras
+                Bundle dataBundle = new Bundle();
+                dataBundle.putLong(Training_DAO.TABLE_NAME + Training_DAO.COLUMN_ID, trainingList.get(currentPos).getId());
+                //add data
+                intent.putExtras(dataBundle);
+                //start activity
+                startActivityForResult(intent, 0);
+                return true;
+
+            case R.id.Delete:
                 mDialog = AlertToDelete_DialogFragment.newInstance();
                 mDialog.show(mFragmentManager, "Alert");
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
+        }
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        if (requestCode == 0) {
+            try {
+                trainingList = training_dao.getAll();
+                mListTrainings.setTrainingList(getNamesList(trainingList));
+                mListTrainings.updateList();
+
+            } catch (GenericDAOException e) {
+                e.printStackTrace();
+            }
+            mDetailsTraining.clearDetails();
+            currentPos = -1;
+            mOptionsMenu.findItem(R.id.Delete).setVisible(false);
+            mOptionsMenu.findItem(R.id.Edit).setVisible(false);
         }
     }
 }
